Simple prototype on how to log command arguments
